requirements.txt
sympy
rdkit-pypi
